Fix a bug in 'Lexer#SkipWhile'
Modified implementations of 'Lexer#SkipWhile'. The method must return a
value of skipping count. But the method had been returning -1 if it
reaches the end of input file. So in parent method 'Lexer#Tokenize', the
control fall into an infinite loop if there are whitespaces just before
the end of input file.